refactor: decouple entity graph chart component from hash-frontend
